Moving AudienceMetadataParser to gcp auth filter and better error handling
