+++ xds/src/main/java/io/grpc/xds/ClusterMetadataRegistry.java
@@ -18,7 +18,7 @@ package io.grpc.xds;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.InvalidProtocolBufferException;
-import io.envoyproxy.envoy.extensions.filters.http.gcp_authn.v3.Audience;
+import io.grpc.xds.GcpAuthenticationFilter.AudienceMetadataParser;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -34,11 +34,7 @@ final class ClusterMetadataRegistry {
   private final Map<String, ClusterMetadataValueParser> supportedParsers = new HashMap<>();
 
   private ClusterMetadataRegistry() {
-    registerParsers(
-        new Object[][]{
-            {"extensions.filters.http.gcp_authn.v3.Audience", new AudienceMetadataParser()},
-            // Add more parsers here as needed
-        });
+    registerParser(new AudienceMetadataParser());
   }
 
   static ClusterMetadataRegistry getInstance() {
@@ -49,16 +45,14 @@ final class ClusterMetadataRegistry {
     return supportedParsers.get(typeUrl);
   }
 
-  private void registerParsers(Object[][] parserEntries) {
-    for (Object[] entry : parserEntries) {
-      String typeUrl = (String) entry[0];
-      ClusterMetadataValueParser parser = (ClusterMetadataValueParser) entry[1];
-      supportedParsers.put(typeUrl, parser);
-    }
+  private void registerParser(ClusterMetadataValueParser parser) {
+    supportedParsers.put(parser.getTypeUrl(), parser);
   }
 
-  @FunctionalInterface
   interface ClusterMetadataValueParser {
+
+    String getTypeUrl();
+
     /**
      * Parses the given {@link Any} object into a specific metadata value.
      *
@@ -67,24 +61,5 @@ final class ClusterMetadataRegistry {
      * @throws InvalidProtocolBufferException if the parsing fails.
      */
     Object parse(Any any) throws InvalidProtocolBufferException;
-  }
-
-  /**
-   * Parser for Audience metadata type.
-   */
-  class AudienceMetadataParser implements ClusterMetadataValueParser {
-    @Override
-    public String parse(Any any) throws InvalidProtocolBufferException {
-      if (any.is(Audience.class)) {
-        Audience audience = any.unpack(Audience.class);
-        String url = audience.getUrl();
-        if (url.isEmpty()) {
-          throw new InvalidProtocolBufferException("Audience URL is empty.");
-        }
-        return url;
-      } else {
-        throw new InvalidProtocolBufferException("Unexpected message type: " + any.getTypeUrl());
-      }
-    }
   }
 }
